* extract slice element container into a file
